added controllers

+++ backend/controllers/boardingController.js
@@ -2,6 +2,7 @@ import validator from "validator";
 
 import User from "../models/userModel.js";
 import Boarding from "../models/boardingModel.js";
+import mongoose from "mongoose";
 
 //handle images
 
@@ -41,15 +42,21 @@ export const createBoarding = async (req, res) => {
 
 export const deleteBoarding = async (req,res) => {
     try {
+
+        const id = req.params.id;
         
-        const boarding = await Boarding.findById(req.params.id);
+        if(!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({error: "Incorrect ID format"});  
+        }
+
+        const boarding = await Boarding.findById(id);
 
         if(!boarding){
             return res.status(401).json({error: 'Boarding not found to delete'})
         }
 
         if(boarding.userId.toString() !== req.user._id.toString()){
-            return res.status(401).json({ error: 'You are not authorized to delete this post'});
+            return res.status(401).json({ error: 'You are not authorized to delete this boarding'});
         }
 
         await Boarding.findByIdAndDelete(req.params.id);
@@ -76,6 +83,63 @@ export const getAllBoardings = async (req,res) =>{
 
     } catch (error) {
         console.log('error in getAllBoardings controller', error);
+        res.status(500).json({error: 'Internal Server Error'})
+    }
+}
+
+export const getUserBoarding = async (req,res) => {
+    try {
+        
+        const id = req.params.id;
+
+        if(!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({error: "Incorrect ID format"});  
+        }
+
+        if(id !== req.user._id.toString()){
+            return res.status(401).json({ error: 'You are not authorized to view'});
+        }
+
+        const boardings = await Boarding.find({ userId: id });
+
+        if (boardings.length === 0){
+            return res.status(200).json([]);
+        }
+
+        res.status(200).json(boardings);
+
+    } catch (error) {
+        console.log('error in getUserBoardings controller', error);
+        res.status(500).json({error: 'Internal Server Error'})
+    }
+}
+
+export const updateBoarding = async (req,res) =>{
+    try {
+        
+        const id = req.params.id;
+        
+        if(!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({error: "Incorrect ID format"});  
+        }
+
+        const boarding = await Boarding.findById(id);
+
+        if(!boarding){
+            return res.status(401).json({error: 'Boarding not found to update'})
+        }
+
+        if(boarding.userId.toString() !== req.user._id.toString()){
+            return res.status(401).json({ error: 'You are not authorized to update this boarding'});
+        }
+
+        const updatedBoarding = await Boarding.findOneAndUpdate({_id: id}, {...req.body});
+
+        res.status(200).json({updatedBoarding})
+
+
+    } catch (error) {
+        console.log('error in deleteBoarding controller', error);
         res.status(500).json({error: 'Internal Server Error'})
     }
 }
